completed task 2 from bonus

+++ routes/web.php
@@ -19,6 +19,6 @@ Route::get('/', [TaskController::class, 'index'])->name('tasks.index');
 Route::get('/tasks/create', [TaskController::class, 'create'])->name('tasks.create');
 Route::post('/tasks', [TaskController::class, 'store'])->name('tasks.store');
 Route::patch('/tasks/toggle/{task}', [TaskController::class, 'toggleCompleted'])->name('tasks.toggle');
-Route::delete('/tasks/{task}', [TaskController::class, 'destroy'])->name('tasks.destroy');
+Route::delete('/tasks/{task}', [TaskController::class, 'destroy'])->name('tasks.destroy')->middleware('admin');
 Route::get('/tasks/edit/{task}', [TaskController::class, 'edit'])->name('tasks.edit');
 Route::patch('/tasks/{task}', [TaskController::class, 'update'])->name('tasks.update');
